+++ changelog.md
@@ -1,6 +1,15 @@
+# v0.2.0-alpha.1
+
+- Color settings for transparent lists.
+- Use new method for transparent lists colors (could potentially crash with geode updates).
+- Don't shorten download count setting.
+- Updated settings.
+- Updated changelog.
+- Fix the fix mod info size setting.
+
 # v0.1.0
 
-- Transparent Lists
-- List Background Transparency
-- Larger Logos
-- Fix Mod Info Size
+- Added transparent lists setting.
+- Added transparent lists background transparency setting.
+- Added larger logos setting.
+- Added fix mod info size setting.